Deleted LazyLoading from entity. Minor edit to the vaultmodule to include data from other tables.
PostMeasurementView.xaml - changed the scrollview to a listview with bindings. Used ThumbnailView as datasource.
PostMeasurementViewModel - changed the entire class to work with thumbnails in stead of a single source of data.

Note: work in progress, most code is copied from the vault edit classes since this one does pretty much the same. Should give no errors but the code does not do anything useful yet.